+++ console/src/components/Services/Data/PermissionsSummary/PermissionsSummary.scss
@@ -1,4 +1,4 @@
-@import "../../Common/Common.scss";
+@import "../../../Common/Common";
 
 //.rolesTableWrapper {
 //}
